+++ models/user.js
@@ -1,6 +1,5 @@
 const mongoose = require("mongoose");
 const jwt = require("jsonwebtoken");
-const Joi = require("joi");
 
 const userSchema = new mongoose.Schema({
   firstName: { type: String, minLength: 3, maxLength: 55, required: true },
@@ -28,17 +27,4 @@ userSchema.methods.generateAuthToken = function () {
 
 const User = mongoose.model("User", userSchema);
 
-function validateUser(user) {
-  const schema = Joi.object({
-    firstName: Joi.string().min(3).required(),
-    lastName: Joi.string().min(3).required(),
-    email: Joi.string().min(5).required().email(),
-    password: Joi.string().min(8).max(255).required(),
-    isManager: Joi.boolean(),
-  });
-
-  return schema.validate(user);
-}
-
 exports.User = User;
-exports.validate = validateUser;
