+++ startup/routes.js
@@ -1,8 +1,7 @@
+require("express-async-errors");
 const express = require("express");
 const signup = require("../routes/signup");
 const login = require("../routes/login");
-
-require("express-async-errors");
 
 module.exports = function (app) {
   app.use(express.json());
